AGENTES IA/ABC_ licitaciones/requirements.txt
# Agente de Contrataciones Públicas - Perú
# Dependencias del proyecto

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# OpenAI
openai==1.12.0

# Vector Database & RAG
chromadb==0.4.22
langchain==0.1.6
langchain-google-genai==0.0.9
google-generativeai==0.3.2
langchain-community==0.0.20
pypdf==4.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Production
gunicorn==21.2.0
